wheezy-stage1: chmod authorized_keys file
On systems where newly created files defaulted to rw group permissions (such
as those where umask is 0002) dropbear would refuse to authenticate ssh
logins, complaining that the file should be writable only by the owning user.
Therefore I explicitly chmod it to 600 (and of course, it could reasonably
have been 644 but given the single user nature of the environment it doesn't
really matter).