wheezy-stage3: fix setting of root password
Previously, the root password was simply deleted. This allowed a local user to
log in as root with no password. SSH wouldn't allow root to log in, but having
a default root login is not desirable. Set encrypted root password to *
instead, which is what was originally intended. root login is now disabled
locally and remotely.