wheezy-stage2,common: use qemu nameserver during image build
Previously the Google nameserver at 8.8.8.8 was used on the chrooted image
during the build process. However, I had reports of a user who was unable to
use this server, and it seems sensible to rely purely on the DNS setup of the
build host, so use 10.0.2.3 (the host qemu) during the whole build process. We
still default /etc/resolv.conf to 8.8.8.8 upon image export.